Skeleton for logging
We will need to figure out exactly what our "traders" are doing
in order to improve them. Logging will be the first part of this
long and arduous iteration process. In the future, we will use
those same logs to construct charts for visualising the evolution
of the market and how the traders respond.

Testing:
None for now